Adding types file to program package
